;doc: cli: split journal syntax into "main" and "other"
This comes from a review of journal syntax, in preparation for
supporting more Ledger syntax.

To reduce overload for learners, I have trimmed the "main" feature
list, and de-emphasised a number of my not-so-favourite features by
moving them under "Other syntax", with downsides noted for good
measure. Secondary dates, balance assignments, and auto postings
are some of the de-emphasised features.